Add getEvents and restartPrimaries operations.
getEvents retrives the project-level events for a given project.
restartPrimaries does exactly that - its intended usecase it to test
Mongo failover and election scenarioes.

Some tests are added. The most reliable way to test a client for an
API is to test against an actual instance of that API, so the tests
are skipped by default and run only if certain environment variables
are set. I have tested both new endpoints against a Clever project.

API version 0.6.3 -> 0.7.0

Some changes in gen-go are due to the newer version of wag. Nothing
breaking there.